Initial import of Chicago Fire Model code and data
This is a top-skim tree from Uptake.org, received from them on
2017-11-07 as red_cross-master@9e8447b4126.zip -- presumably
9e8447b4126 is the git commit ID from their internal repository.

The code is intended by all parties to be released under an open
source license with the Red Cross as copyright holder, as per
agreement on or about that date between Uptake Technologies Inc and
the American Red Cross of Chicago & Northern Illinois.  This commit
reflects the code tree exactly as received from Uptake; the open
source license will be applied in a subsequent commit.